Refactor studio and explorer views
